+++ packages/account/src/user/infrastructure/infrastructure.module.ts
@@ -2,13 +2,13 @@ import { DatabaseModule, DatabaseSchema } from "@core/database";
 import { Module, Provider } from "@nestjs/common";
 import { ConfigModule, ConfigService } from "@nestjs/config";
 import { UserRepositoryPort } from "../domain/port/user-repository.port";
-import { UserOrmRepository } from "./repository/user";
+import { UserOrmRepository } from "./repository/user.orm.repository";
 import { UserController } from "./presentation/user/user.controller";
 import { PetController } from "../../pet/infrastructure/presentation/pet/pet.controller";
 import { StorageModule } from "@core/storage";
 import { AuthInternalApiAdapter } from "./adapter/auth-internal-api.adpater";
 import { AuthInternalApiPort } from "../domain/port/auth-internal-api.port";
-import { AuthOrmRepository } from "@app/auth/infrastructure/repository/auth.repository";
+import { AuthOrmRepository } from "@app/auth/infrastructure/repository/auth.orm.repository";
 import { AuthMapper } from "@app/auth/infrastructure/repository/mapper/auth.mapper";
 import { UserStoragePort } from "../domain/port/user-storage.port";
 import { UserStorageAdapter } from "./adapter/user-storage.adapter";
